+++ processor/src/fast/join.rs
@@ -6,7 +6,7 @@ use miden_core::mast::{JoinNode, MastForest, MastNodeId};
 use crate::{
     AsyncHost,
     continuation_stack::{Continuation, ContinuationStack},
-    fast::{BreakReason, FastProcessor, Tracer, step::Stopper, trace_state::NodeExecutionState},
+    fast::{BreakReason, FastProcessor, Tracer, step::Stopper},
 };
 
 impl FastProcessor {
@@ -24,7 +24,7 @@ impl FastProcessor {
     ) -> ControlFlow<BreakReason> {
         tracer.start_clock_cycle(
             self,
-            NodeExecutionState::Start(node_id),
+            Continuation::StartNode(node_id),
             continuation_stack,
             current_forest,
         );
@@ -54,7 +54,7 @@ impl FastProcessor {
     ) -> ControlFlow<BreakReason> {
         tracer.start_clock_cycle(
             self,
-            NodeExecutionState::End(node_id),
+            Continuation::FinishJoin(node_id),
             continuation_stack,
             current_forest,
         );